+++ demo/src/main/java/cn/bingoogolapple/swipebacklayout/demo/App.java
@@ -2,12 +2,12 @@ package cn.bingoogolapple.swipebacklayout.demo;
 
 import android.app.Application;
 
-import cn.bingoogolapple.swipebacklayout.BGASwipeBackManager;
+import cn.bingoogolapple.swipebacklayout.BGASwipeBackHelper;
 
 /**
  * 作者:王浩 邮件:bingoogolapple@gmail.com
  * 创建时间:16/12/28 下午11:55
- * 描述:「必须在 Application 的 onCreate 方法中执行 BGASwipeBackManager.getInstance().init(this) 来初始化滑动返回」
+ * 描述:「必须在 Application 的 onCreate 方法中执行 BGASwipeBackHelper.init(this) 来初始化滑动返回」
  */
 public class App extends Application {
 
@@ -15,7 +15,7 @@ public class App extends Application {
     public void onCreate() {
         super.onCreate();
 
-        // 必须在 Application 的 onCreate 方法中执行 BGASwipeBackManager.getInstance().init(this) 来初始化滑动返回
-        BGASwipeBackManager.getInstance().init(this);
+        // 必须在 Application 的 onCreate 方法中执行 BGASwipeBackHelper.init(this) 来初始化滑动返回
+        BGASwipeBackHelper.init(this);
     }
 }