+++ library/src/main/java/cn/bingoogolapple/swipebacklayout/BGASwipeBackHelper.java
@@ -17,6 +17,7 @@
 package cn.bingoogolapple.swipebacklayout;
 
 import android.app.Activity;
+import android.app.Application;
 import android.content.Intent;
 import android.support.annotation.DrawableRes;
 import android.support.annotation.FloatRange;
@@ -31,6 +32,15 @@ public class BGASwipeBackHelper {
     private Activity mActivity;
     private Delegate mDelegate;
     private BGASwipeBackLayout mSwipeBackLayout;
+
+    /**
+     * 必须在 Application 的 onCreate 方法中调用
+     *
+     * @param application
+     */
+    public static void init(Application application) {
+        BGASwipeBackManager.getInstance().init(application);
+    }
 
     public BGASwipeBackHelper(Activity activity, Delegate delegate) {
         mActivity = activity;